Move more generic utils into lib/
