+++ packages/liveblocks-core/src/lib/__tests__/utils.test.ts
@@ -1,0 +1,17 @@
+import { compact } from "../utils";
+
+describe("compact", () => {
+  it("compact w/ empty list", () => {
+    expect(compact([])).toEqual([]);
+  });
+
+  it("compact removes nulls and undefined values", () => {
+    expect(compact(["a", "b", "c"])).toEqual(["a", "b", "c"]);
+    expect(compact(["x", undefined])).toEqual(["x"]);
+    expect(compact([0, null, undefined, NaN, Infinity])).toEqual([
+      0,
+      NaN,
+      Infinity,
+    ]);
+  });
+});
