+++ packages/liveblocks-core/src/lib/freeze.ts
@@ -1,0 +1,8 @@
+/**
+ * Freezes the given argument, but only in development builds. In production
+ * builds, this is a no-op for performance reasons.
+ */
+export const freeze: typeof Object.freeze =
+  process.env.NODE_ENV === "production"
+    ? (((x: unknown) => x) as typeof Object.freeze)
+    : Object.freeze;
