+++ packages/liveblocks-core/src/lib/utils.ts
@@ -1,0 +1,33 @@
+export function remove<T>(array: T[], item: T): void {
+  for (let i = 0; i < array.length; i++) {
+    if (array[i] === item) {
+      array.splice(i, 1);
+      break;
+    }
+  }
+}
+
+/**
+ * Removes null and undefined values from the array, and reflects this in the
+ * output type.
+ */
+export function compact<T>(items: readonly T[]): NonNullable<T>[] {
+  return items.filter(
+    (item: T): item is NonNullable<T> => item !== null && item !== undefined
+  );
+}
+
+/**
+ * Returns a new object instance where all explictly-undefined values are
+ * removed.
+ */
+export function compactObject<O>(obj: O): O {
+  const newObj = { ...obj };
+  Object.keys(obj).forEach((k) => {
+    const key = k as keyof O;
+    if (newObj[key] === undefined) {
+      delete newObj[key];
+    }
+  });
+  return newObj;
+}
