+++ packages/liveblocks-core/src/utils.ts
@@ -23,49 +23,6 @@ import { CrdtType } from "./types";
 import type { CreateOp, Op } from "./types/Op";
 import { OpCode } from "./types/Op";
 
-export function remove<T>(array: T[], item: T): void {
-  for (let i = 0; i < array.length; i++) {
-    if (array[i] === item) {
-      array.splice(i, 1);
-      break;
-    }
-  }
-}
-
-/**
- * Freezes the given argument, but only in development builds. In production
- * builds, this is a no-op for performance reasons.
- */
-export const freeze: typeof Object.freeze =
-  process.env.NODE_ENV === "production"
-    ? (((x: unknown) => x) as typeof Object.freeze)
-    : Object.freeze;
-
-/**
- * Removes null and undefined values from the array, and reflects this in the
- * output type.
- */
-export function compact<T>(items: readonly T[]): NonNullable<T>[] {
-  return items.filter(
-    (item: T): item is NonNullable<T> => item !== null && item !== undefined
-  );
-}
-
-/**
- * Returns a new object instance where all explictly-undefined values are
- * removed.
- */
-export function compactObject<O>(obj: O): O {
-  const newObj = { ...obj };
-  Object.keys(obj).forEach((k) => {
-    const key = k as keyof O;
-    if (newObj[key] === undefined) {
-      delete newObj[key];
-    }
-  });
-  return newObj;
-}
-
 export function creationOpToLiveNode(op: CreateOp): LiveNode {
   return lsonToLiveNode(creationOpToLson(op));
 }
